Update Hipchat support to work w/all versions
The HipChat plugin has changed a good bit in version 0.1.9 such that it
is now purely using the publisher hook and it's acquired a few new
configuration items.  This patch adds support for 0.1.9 and also adds
support for the new custom start and completed messages.

It also changes the behavior when the plugin version data is not
available such that it will assume you're running the latest version.

Fixes SB Bug Id: 2000301 2000324

Change-Id: I39452e877b20bfc5654d46c90e91dd877a8ebeb0